% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubs.R
\name{onfleet_get_hubs}
\alias{onfleet_get_hubs}
\title{Get hubs}
\usage{
onfleet_get_hubs()
}
\description{
Get hubs
}
\seealso{
\url{http://docs.onfleet.com/docs/hubs} \url{http://docs.onfleet.com/docs/entities#hub}
}
