% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{onfleet_get_organizations}
\alias{onfleet_get_organizations}
\title{Get info about organizations}
\usage{
onfleet_get_organizations()
}
\description{
Get info about organizations
}
\seealso{
\url{http://docs.onfleet.com/docs/organizations} \url{http://docs.onfleet.com/docs/entities#organization}
}
