% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrators.R
\name{onfleet_post_admins}
\alias{onfleet_post_admins}
\alias{onfleet_get_admins}
\alias{onfleet_put_admins}
\alias{onfleet_delete_admins}
\title{Manage administrators}
\usage{
onfleet_post_admins(name, email, phone = NULL, isReadOnly = NULL)

onfleet_get_admins()

onfleet_put_admins(id, ...)

onfleet_delete_admins(id)
}
\arguments{
\item{name}{Name of new administrator}

\item{email}{Email of new administrator}

\item{phone}{Phone number of new administrator}

\item{isReadOnly}{Indicate whether admin is read-only}

\item{id}{ID of administrator}

\item{...}{values to update}
}
\description{
Manage administrators
}
\seealso{
\url{http://docs.onfleet.com/docs/administrators} \url{http://docs.onfleet.com/docs/entities#administrator}
}
