% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{onfleet_post_recipients}
\alias{onfleet_post_recipients}
\alias{onfleet_put_recipients}
\alias{onfleet_find_recipients}
\alias{onfleet_get_recipient}
\title{Manage Recipients}
\usage{
onfleet_post_recipients(name, phone, notes = NULL,
  skipSMSNotifications = NULL, skipPhoneNumberValidation = NULL)

onfleet_put_recipients(id, ...)

onfleet_find_recipients(name = NULL, phone = NULL)

onfleet_get_recipient(id)
}
\arguments{
\item{name}{Recipient name}

\item{phone}{Recipient phone number}

\item{notes}{Notes}

\item{skipSMSNotifications}{Indiate if recipient should not receive notifications}

\item{skipPhoneNumberValidation}{Indicate if you would like API not to validate phone number}

\item{id}{ID of recipients}

\item{...}{values to update}
}
\description{
Manage Recipients
}
\seealso{
\url{http://docs.onfleet.com/docs/recipients} \url{http://docs.onfleet.com/docs/entities#recipient}
}
