% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{onfleet_post_tasks}
\alias{onfleet_post_tasks}
\alias{onfleet_get_tasks}
\alias{onfleet_get_task}
\alias{onfleet_put_task}
\alias{onfleet_post_complete_task}
\alias{onfleet_delete_task}
\title{Manage Tasks}
\usage{
onfleet_post_tasks(merchant, executor, destination, recipients, completeAfter,
  completeBefore, pickupTask, dependencies, notes, autoAssignMode,
  autoAssignTeam, autoAssignMaxAssignedTaskCount,
  autoAssignConsiderDependencies, autoAssignExcludedWorkerIds, containerType,
  containerTeam, containerWorker, quantity, serviceTime)

onfleet_get_tasks(from, to, lastId, state, worker, completeBeforeBefore,
  completeAfterAfter)

onfleet_get_task(id)

onfleet_put_task(id, ...)

onfleet_post_complete_task(id, success, notes)

onfleet_delete_task(id)
}
\arguments{
\item{merchant}{Organization ID of merchant}

\item{executor}{Organization ID of executor}

\item{destination}{ID of destination}

\item{recipients}{Vector of recipient IDs}

\item{completeAfter}{Timestamp}

\item{completeBefore}{Timestamp}

\item{pickupTask}{Indicate if pickup}

\item{dependencies}{Vector of tasks this task depends on}

\item{notes}{Notes}

\item{quantity}{Number of units dropped off}

\item{serviceTime}{Time spent at destination}
}
\description{
Manage Tasks
}
\seealso{
\url{http://docs.onfleet.com/docs/tasks} \url{http://docs.onfleet.com/docs/entities#task}
}
