% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workers.R
\name{onfleet_post_workers}
\alias{onfleet_post_workers}
\alias{onfleet_get_workers}
\alias{onfleet_get_workers_by_location}
\alias{onfleet_get_worker}
\alias{onfleet_put_workers}
\alias{onfleet_delete_workers}
\alias{onfleet_get_worker_schedule}
\alias{onfleet_put_worker_schedule}
\title{Manage Workers}
\usage{
onfleet_post_workers(name, phone, teams, vehicleType = NULL,
  vehicleDescription = NULL, vehicleLicensePlate = NULL,
  vehicleColor = NULL, capacity = NULL)

onfleet_get_workers(states)

onfleet_get_workers_by_location(long, lat, radius)

onfleet_get_worker(id)

onfleet_put_workers(id, ...)

onfleet_delete_workers(id)

onfleet_get_worker_schedule(id)

onfleet_put_worker_schedule(id, entries)
}
\arguments{
\item{name}{Name of new worker}

\item{phone}{Phone number of new worker}

\item{teams}{Vector of IDs indicating which teams new worker is member of}

\item{vehicleType}{Either CAR, MOTORCYCLE, BICYCLE, or TRUCK}

\item{vehicleDescription}{Description of vehicle}

\item{vehicleLicensePlate}{License Plate}

\item{vehicleColor}{Color}

\item{capacity}{Maximum number of units this worker can carry}

\item{long}{Longtitude}

\item{lat}{Latitude}

\item{radius}{Radius of area around lat/long you want to search for workers}

\item{id}{ID of worker}

\item{...}{values to update}
}
\description{
Manage Workers
}
\seealso{
\url{http://docs.onfleet.com/docs/workers} \url{http://docs.onfleet.com/docs/entities#worker}
}
