% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{onfleet_put_insert_tasks}
\alias{onfleet_put_insert_tasks}
\alias{onfleet_put_container}
\title{Manage Containers}
\usage{
onfleet_put_insert_tasks(containerType, entityId, index, tasks)

onfleet_put_container(tasks)
}
\description{
Manage Containers
}
