% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{onfleet_post_tasks}
\alias{onfleet_post_tasks}
\alias{onfleet_get_tasks}
\alias{onfleet_get_task}
\alias{onfleet_put_task}
\alias{onfleet_post_complete_task}
\alias{onfleet_delete_task}
\title{Manage Tasks}
\usage{
onfleet_post_tasks(merchant = NULL, executor = NULL, destination,
  recipients, completeAfter = NULL, completeBefore = NULL,
  pickupTask = NULL, dependencies = NULL, notes = NULL,
  autoAssignMode = NULL, autoAssignTeam = NULL,
  autoAssignMaxAssignedTaskCount = NULL,
  autoAssignConsiderDependencie = NULLs, autoAssignExcludedWorkerIds = NULL,
  containerType = NULL, containerTeam = NULL, containerWorker = NULL,
  quantity = NULL, serviceTime = NULL)

onfleet_get_tasks(from, to = NULL, lastId = NULL, state = NULL,
  worker = NULL, completeBeforeBefore = NULL, completeAfterAfter = NULL)

onfleet_get_task(id)

onfleet_put_task(id, ...)

onfleet_post_complete_task(id, success, notes)

onfleet_delete_task(id)
}
\arguments{
\item{merchant}{Organization ID of merchant}

\item{executor}{Organization ID of executor}

\item{destination}{ID of destination}

\item{recipients}{Vector of recipient IDs}

\item{completeAfter}{Timestamp}

\item{completeBefore}{Timestamp}

\item{pickupTask}{Indicate if pickup}

\item{dependencies}{Vector of tasks this task depends on}

\item{notes}{Notes}

\item{quantity}{Number of units dropped off}

\item{serviceTime}{Time spent at destination}
}
\description{
Manage Tasks
}
\seealso{
\url{http://docs.onfleet.com/docs/tasks} \url{http://docs.onfleet.com/docs/entities#task}
}
